# Fixes for NonBacktracking NFA mode